git-svn-id: svn://192.168.10.30/autonomos@9 13ecfbb0-6a6f-4a77-9248-64ec958d7be8

+++ src/main/java/com/level2/multiagent/autonomos/Dimiurgos.java
@@ -22,7 +22,7 @@ import com.level2.multiagent.autonomos.scheduler.PlainPopulationScheduler;
 public class Dimiurgos {
 	public static final Logger logger = LoggerFactory.getLogger(Dimiurgos.class);
 	
-	private Integer totalAgentNumber = new Integer(300);
+	private Integer totalAgentNumber = new Integer(10);
 	private ArrayList<IAgent> agents;
 	private Integer iterationsSucceful= new Integer(0);
